+++ registry/service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # service changelog
 
+## 0.1.2 — 2026-07-15
+
+Publish service v0.1.2.
+
+## 0.1.1 — 2026-07-15
+
+Publish service v0.1.1.
+
 ## 0.1.0 — 2026-07-15
 
 Publish service v0.1.0.